Module Down

An OCaml toplevel (REPL) upgrade

See the manual.

Down

val help : unit -> unit

help () prints help about Down.

module Session : sig ... end

Manage sessions.

module History : sig ... end

Manage history.

val tty_no_faint : unit -> unit

tty_no_faint () disables uses of ANSI faint by down. Some terminals switch to different colors which may be unpleasant, call this function if that is the case.

module Private : sig ... end

Private.