Module Down.Private

Private.

Do not use. This is an unstable API subject to change even between minor versions of the library.

module type TOP = sig ... end

OCaml Toplevel API

val set_top : (module TOP) -> unit

set_top t sets the implementation of the OCaml toplevel to t.

val unicode_version : string

unicode_version is the Unicode version on which the Uucp.Break.tty_width_hint data used by down is based.

val tty_test : unit -> unit

tty_test () interactively tests how terminal input is parsed by down.